Do not set the internal name for modules on HP-UX
Using a name that does not end with .sl for the internal name (+h)
on HP-UX may result in a link error for some versions of HP ld.  In
the case of modules, there's no reason to set the internal name so
just omit the "+h name" from module_cmds.